pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "adflux"
version = "0.1.0"
description = "Pure-Python Markdown ↔ Atlassian Document Format (ADF) converter with lossless round-tripping"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Mike Hill", email = "mike@mikejhill.com" }]
maintainers = [{ name = "Mike Hill", email = "mike@mikejhill.com" }]
keywords = [
    "markdown",
    "adf",
    "atlassian",
    "atlassian-document-format",
    "commonmark",
    "gfm",
    "converter",
    "document-conversion",
    "confluence",
    "jira",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Documentation",
    "Topic :: Software Development :: Documentation",
    "Topic :: Text Processing :: Markup",
    "Topic :: Text Processing :: Markup :: Markdown",
    "Topic :: Utilities",
    "Typing :: Typed",
]
dependencies = [
    "markdown-it-py>=3.0",
    "mdit-py-plugins>=0.4",
    "panflute>=2.3.1",
    "PyYAML>=6.0",
    "jsonschema>=4.21",
    "typer>=0.12",
    "structlog>=24.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "hypothesis>=6.100",
    "ruff>=0.5",
    "mypy>=1.10",
    "types-PyYAML",
    "types-jsonschema",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "poethepoet>=0.27",
]

[project.scripts]
adflux = "adflux.cli:app"

[project.urls]
Homepage = "https://github.com/mikejhill/adflux"
Documentation = "https://github.com/mikejhill/adflux/tree/main/docs"
Repository = "https://github.com/mikejhill/adflux"
Issues = "https://github.com/mikejhill/adflux/issues"
Changelog = "https://github.com/mikejhill/adflux/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["src/adflux"]

[tool.hatch.build.targets.sdist]
include = [
    "src/**/*.py",
    "src/**/*.yaml",
    "src/**/*.json",
    "tests/**/*.py",
    "tests/**/*.json",
    "examples/**/*",
    "docs/**/*",
    "README.md",
    "CHANGELOG.md",
    "LICENSE",
    "pyproject.toml",
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra --strict-markers --strict-config"
testpaths = ["tests"]
pythonpath = ["src", "."]
markers = [
    "slow: marks tests as slow",
    "roundtrip: round-trip fidelity tests",
    "e2e: end-to-end tests that hit a live Confluence Cloud site",
]

[tool.coverage.run]
branch = true
source = ["src/adflux"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "\\.\\.\\.",
]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "W",    # pycodestyle
    "F",         # pyflakes
    "I",         # isort
    "B",         # bugbear
    "UP",        # pyupgrade
    "C4",        # comprehensions
    "SIM",       # simplify
    "RET",       # returns
    "PTH",       # pathlib
    "RUF",       # ruff-specific
    "N",         # pep8-naming
    "D",         # pydocstyle
]
ignore = [
    "D100", "D104", "D105", "D107",  # allow missing docstrings on modules/packages/magic/__init__
    "D203", "D213",                   # conflicting with D211/D212
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["D"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
packages = ["adflux"]
mypy_path = "src"

[[tool.mypy.overrides]]
module = ["panflute.*", "markdown_it.*", "mdit_py_plugins.*"]
ignore_missing_imports = true

[tool.poe.tasks]
_format       = "ruff format src tests"
_format_check = "ruff format --check src tests"
_lint         = "ruff check src tests"
_lint_fix     = "ruff check --fix src tests"
_typecheck    = "mypy"

format       = ["_format", "_lint_fix"]
lint         = ["_format_check", "_lint"]
typecheck    = ["_typecheck"]
test         = "pytest --ignore=tests/e2e"
test-all     = "pytest"
test-e2e     = "pytest tests/e2e -v"
cov          = "pytest --ignore=tests/e2e --cov=adflux --cov-report=term-missing"
build        = "python -m build"
check        = ["lint", "typecheck", "test"]

[tool.poe.tasks.clean]
help = "Remove build artifacts and caches."
shell = """
python -c "
import shutil, pathlib
for p in ['dist','build','.pytest_cache','.mypy_cache','.ruff_cache','htmlcov','.coverage']:
    pp = pathlib.Path(p)
    if pp.is_dir(): shutil.rmtree(pp, ignore_errors=True)
    elif pp.exists(): pp.unlink()
for egg in pathlib.Path('.').glob('src/*.egg-info'): shutil.rmtree(egg, ignore_errors=True)
for egg in pathlib.Path('.').glob('*.egg-info'): shutil.rmtree(egg, ignore_errors=True)
"
"""
